+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Features
 * [#30] - Introduce charts metadata
 * [#37] - Enhance a view metadata fields
+* [#38] - Introduce a query API with a proper filtering which will fetch an initial content
 
 ### Improvements
 * [#35] - Remove redundant cardinality field for a view metadata